Containerize the process of publishing packages to fake registry
