+++ packages/e2e-tests/Dockerfile.publish-packages
@@ -1,0 +1,6 @@
+# This Dockerfile exists for the purpose of using a specific node/npm version (ie. the same we use in CI) to run npm publish with
+ARG NODE_VERSION=16.15.1
+FROM node:${NODE_VERSION}
+
+WORKDIR /sentry-javascript/packages/e2e-tests
+CMD [ "yarn", "ts-node", "publish-packages.ts" ]
