+++ packages/e2e-tests/test-registry.npmrc
@@ -1,2 +1,3 @@
-registry=http://localhost:4873/
+@sentry:registry=http://localhost:4873
+@sentry-internal:registry=http://localhost:4873
 //localhost:4873/:_authToken=some-token
